promptnova/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
requests==2.32.5
groq>=1.0.0
python-dotenv==1.0.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.3.2
PyYAML==6.0.1
